% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignPosterior.R
\name{alignPosterior}
\alias{alignPosterior}
\title{alignPosterior}
\usage{
alignPosterior(hM)
}
\description{
Aligns posterior in terms of variable susceptible for label switching
}
\examples{
# Align the posterior for a previously fitted HMSC model
m = alignPosterior(TD$m)

}
