% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.HMSCdata.R, R/as.HMSCparam.R,
%   R/as.HMSCprior.R
\name{as.HMSCdata}
\alias{as.HMSCdata}
\alias{as.HMSCparam}
\alias{as.HMSCprior}
\title{Format data, parameters, and priors to "HMSC" classes}
\usage{
as.HMSCdata(Y = NULL, X = NULL, Tr = NULL, Phylo = NULL, Auto = NULL,
  Random = NULL, scaleX = TRUE, scaleTr = TRUE, interceptX = TRUE,
  interceptTr = TRUE)

as.HMSCparam(data, priors, varDist = NULL, residVar = NULL, paramX = NULL,
  meansParamX = NULL, varX = NULL, paramTr = NULL, paramPhylo = NULL,
  paramAuto = NULL, latentAuto = NULL, paramLatentAuto = NULL,
  shrinkLocalAuto = NULL, paramShrinkGlobalAuto = NULL, latent = NULL,
  paramLatent = NULL, shrinkLocal = NULL, paramShrinkGlobal = NULL,
  nsp = NULL)

as.HMSCprior(data, family = "probit", varDistShape = NULL,
  varDistScale = NULL, varXDf = NULL, varXScaleMat = NULL,
  meansParamX = NULL, varMeansParamX = NULL, residVar = NULL,
  paramTr = NULL, varTr = NULL, paramPhylo = NULL, paramAutoDist = NULL,
  paramAutoWeight = NULL, shrinkOverall = NULL, shrinkSpeed = NULL,
  shrinkLocal = NULL)
}
\arguments{
\item{Y}{A matrix or a data.frame where each column is a species.}

\item{X}{A matrix or a data.frame where each column is a descriptor of the sites.}

\item{Tr}{A matrix or a data.frame where each row is a trait characterizing all species.}

\item{Phylo}{A square correlation or covariance matrix describing the phylogenetic correlation (or covariance) between pairs of all species (see details).}

\item{Auto}{A data.frame or a list of matrices presenting the spatial (or temporal) coordinates of the samples (see details).}

\item{Random}{A factor defining a random effect on the samples or a data.frame characterizing multiple random effects.}

\item{scaleX}{Logical. Whether the columns of X should be centred and divided by the standard deviation. Default is TRUE.}

\item{scaleTr}{Logical. Whether the rows of Tr should be centred and divided by the standard deviation. Default is TRUE.}

\item{interceptX}{Logical. Whether a column of 1s should be added to X. Default is TRUE.}

\item{interceptTr}{Logical. Whether a row of 1s should be added to Tr. Default is TRUE.}

\item{data}{An object of the class \code{HMSCdata}.}

\item{priors}{An object of the class \code{HMSCprior}.}

\item{varDist}{Vector of parameters with as many entries as there are species. Each values in this vector describes how much variation is not explained by the model.}

\item{residVar}{Vector of parameters with as many entries as there are species. Each values in this vector describes how much variation is not explained by the model.}

\item{paramX}{A matrix or data.frame of model parameters defining how species (rows) are characterized by the descriptors (columns).}

\item{meansParamX}{A vector of parameters defining how an average species reacts to each descriptor in X. In as.HMSCprior, this argument defines the prior information associated to parameters of the same name.}

\item{varX}{Symmetric covariance matrix. Each dimension of this matrix should be equal to the number of explanatory variables.}

\item{paramTr}{In function as.HMSCparam, it is a matrix of model parameters defining how descriptors (rows) characterizes traits (columns). In function as.HMSCprior, it is a matrix of prior defining how descriptors (rows) characterizes traits (columns).}

\item{paramPhylo}{Numeric. In function as.HMSCparam, it defines the importance of phylogeny in structuring the community. In as.HMSCprior, this argument defines prior information associated to the parameter of the same name.}

\item{paramAuto}{Vector. In function as.HMSCparam, it is a vector of numerical values defining the autocorrelated patterns for each factor in \code{Auto}. These values can range from 0 to the largest distance between samples in the autocorrelated level. In as.HMSCprior, this argument defines prior information associated to the parameter of the same name.}

\item{latentAuto}{List. Each level of the list includes a matrix of autocorrelated latent (unsampled) variables. There are as many level in the list as there are columns in \code{Auto}. The number of rows in each matrix corresponds to the number of levels in each autocorrelated factor in \code{Auto}.}

\item{paramLatentAuto}{List. Each level of the list includes a matrix of parameters for the autocorrelated latent (unsampled) variables. There are as many level in the list as there are columns in \code{Auto}. Each matrix has as many rows as there are species in \code{Y} and as many columns as there are autocorrelated latent variables in \code{latentAuto}.}

\item{shrinkLocalAuto}{List. Each level of the list includes a matrix of values describing how much shrinkage should be imposed on each parameter of the autocorrelated latent (unsampled) variables (\code{paramLatentAuto}). There are as many level in the list as there are columns in \code{Auto}. The size of each matrix is the same as the size of \code{paramLatentAuto}.}

\item{paramShrinkGlobalAuto}{List. Each level of the list includes a vector of values describing how much shrinkage should be imposed on each autocorrelated latent variables (\code{latent}) as a whole. There are as many level in the list as there are columns in \code{Auto}. The number of elements in each vector equals the number of variables in \code{latentAuto}.}

\item{latent}{List. Each level of the list includes a matrix of latent (unsampled) variables. There are as many level in the list as there are columns in \code{Random}. The number of rows in each matrix corresponds to the number of levels in each factor in \code{Random}.}

\item{paramLatent}{List. Each level of the list includes a matrix of parameters for the latent (unsampled) variables. There are as many level in the list as there are columns in \code{Random}. Each matrix has as many rows as there are species in \code{Y} and as many columns as there are latent variables in \code{latent}.}

\item{shrinkLocal}{Numeric. Hyperparameter of a gamma distribution defining the local shrinkage for latent variables.}

\item{paramShrinkGlobal}{List. Each level of the list includes a vector of values describing how much shrinkage should be imposed on each latent variables (\code{latent}) as a whole. There are as many level in the list as there are columns in \code{Random}. The number of elements in each vector equals the number of variables in \code{latent}.}

\item{nsp}{Numeric. Number of species in \code{Y}.}

\item{family}{Character string defining the type of generalized regression model for which priors need to be defined.}

\item{varDistShape}{Numeric. Prior for the variance of the normal distribution. It is the shape parameter of a gamma distribution from which the variance is sampled.}

\item{varDistScale}{Numeric. Prior for the variance of the normal distribution. It is the scale parameter of a gamma distribution from which the variance is sampled.}

\item{varXDf}{Numeric. Prior for varX. It is the number of degrees of freedom of the inverse Wishart distribution from which varX is sampled.}

\item{varXScaleMat}{Symmetric covariance matrix. Prior for varX. It is the scale matrix of the inverse Wishart distribution from which varX is sampled.}

\item{varMeansParamX}{In as.HMSCprior, it is a square matrix of parameters defining how \code{meansParamX} varies. This argument is only used when \code{Phylo} is included in the analysis.}

\item{varTr}{In function as.HMSCdata, it is a symmetric covariance matrix. Each dimension of this matrix should be equal to the number of traits. In function as.HMSCprior, it is a symmetric covariance matrix of prior. Each dimension of this matrix should be equal to the number of traits.}

\item{paramAutoDist}{Prior. Matrix with a single column defining the potential values given in \code{paramAuto} can take. The size of \code{priorParamAutoGrid} needs to be equal to the size of \code{priorParamAutoWeight}.}

\item{paramAutoWeight}{Prior. Matrix with a single column defining the importance each value in \code{priorParamAutoDist} may have. The size of \code{priorParamAutoWeight} needs to be equal to the size of \code{priorParamAutoGrid}.}

\item{shrinkOverall}{A vector of length 2 defining the shape and scale hyperparameter of a gamma distribution. The size of the first parameters indicate how the overall shrinkage of the autocorrelated latent variable is handled (see details).}

\item{shrinkSpeed}{A vector of length 2 defining the shape and scale hyperparameter of a gamma distribution. The size of the first parameters indicate how the speed of the shrinkage of the autocorrelated latent variable is handled (see details).}
}
\value{
An object of the class \code{HMSCdata} is returned by \code{as.HMSCdata}.
An object of the class \code{HMSCparam} is returned by \code{as.HMSCparam}. 
An object of the class \code{HMSCprior} is returned by \code{as.HMSCprior}. 

All of these objects present the same output data as the input except for \code{as.HMSCparam} where \code{meansparamX} is also defined.
}
\description{
These functions are used to format the data, parameters, and priors to object of class "HMSC".
}
\details{
Although the structure of the \code{Auto} argument may seem unusual as a list of data.frames, this structures give the possibility to account for different type of autocorrelation structure simultaneously (e.g. spatial and temporal) in the same analysis. Although it is usual for autocorrelation to be accounted for by either 1 or 2 dimensions (so 1 or 2 columns), the analyses carried out in this package can account for autocorrelation at any number of dimensions.

Functions \code{as.HMSCparam} and \code{as.HMSCprior} do not need to be used when estimating a model using \code{\link{hmsc}}. In \code{\link{hmsc}}, if \code{as.HMSCparam} and \code{as.HMSCprior} are not defined, the function will define a set of default parameters and priors. \code{as.HMSCparam} and \code{as.HMSCprior} were designed so that some, not necessarily all, priors and parameters need to be defined. For example, if one defines only the prior (parameters) for \code{varTr}, all other priors (parameters) will be defined with default priors (parameter). 

In \code{as.HMSCparam}, the argument \code{family} is used to define the  \code{paramX} by default using a univariate generalized linear model for each species. Currently, only \code{binomial(link = "probit")} should be used because \code{\link{hmsc}} only estimate probit models. 

In the hmsc function, the \code{Phylo} argument has to be a square correlation matrix. So, if a covariance matrix is given as \code{Phylo} the \code{as.HMSCdata} function will convert it to a correlation matrix.

The choosing the values for \code{shrinkOverall} and \code{shrinkSpeed} is not straightforward. Bhattacharya and Dunson (2011) proposed to use a 2.1 and 1 for \code{shrinkOverall} and 3.1 and 1 for \code{shrinkSpeed}. However, for presence-absence data, it seems that these shrinkage values (especially the scale [first] parameter) are not large enough, which leads to a too many latent variables defined. It is the reason why we suggest to use larger scale values. As such, if they are not predefine, the scale value for \code{shrinkOverall} is 10 whereas the scale value for \code{shrinkSpeed} is 15. These two scale parameters are likely the only one that need to be chosen more carefully to ensure that the model estimation is accurate.

\code{shrinkOverall} and \code{shrinkSpeed} are arguments that are used both for the latent variables as well as the autocorrelated latent variables. 

When predefining the initial value of \code{paramPhylo}, the \code{as.HMSCparam} function may change slightly the value of \code{paramPhylo} to make sure that the \code{hmsc} function runs properly. Since \code{paramPhylo} is estimated using a step function instead of sampling the parameter from a conjugate prior distribution, this step is important to ensure that the code runs well.
}
\examples{

#================
### Generate data
#================
desc <- cbind(1, scale(1:50), scale(1:50)^2)
randomEff <- as.factor(rep(letters[1:5], each=10))
dataBase <- communitySimul(desc, nsp = 30, Random = randomEff)

#=============
### Formatting
#=============
### Format data
formdata <- as.HMSCdata(Y = dataBase$data$Y, X=desc, Random=randomEff, 
						   interceptX = FALSE)

### Format priors
formpriors <- as.HMSCprior(formdata, shrinkOverall = c(50, 1), 
							  shrinkSpeed = c(20, 1))

### Format parameters
formparam <- as.HMSCparam(formdata, formpriors, 
							 paramX = dataBase$param$paramX)

}
\references{
Bhattacharya, A. and Dunson, D.B. (2011) Sparse Bayesian infinite factor models. \emph{Biometrika} \strong{98}, 291--306.
}
\author{
F. Guillaume Blanchet
}
\keyword{classes}
\keyword{manip}
