% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.mcmc.hmsc.R
\name{as.mcmc.hmsc}
\alias{as.mcmc.hmsc}
\title{Convert MCMC parameters estimation to \code{\link{mcmc}} object}
\usage{
\method{as.mcmc}{hmsc}(x, parameters = "paramX", burning = FALSE,
  Random = 1, Auto = 1, ...)
}
\arguments{
\item{x}{An object of the class \code{hmsc}.}

\item{parameters}{A character string defining the parameters for which the convergions needs to be carried out.}

\item{burning}{Logical. Whether the burning iterations should be include (\code{TRUE}) or not (\code{FALSE}).}

\item{\dots}{Addtional arguments passed to \code{\link[coda]{mcmc}}.}
}
\value{
An object of class \code{\link{mcmc}}.
}
\description{
Use the Markov Chain Monte Carlo results obtained from \code{\link{hmsc}} and convert them to an \code{\link{mcmc}} object.
}
\details{
Since the algorithm used adapts the number of latent variables to use, the \code{\link{mcmc}} object associated to \code{paramLatent} only includes information associated to the latent variables that are available for all iterations.
}
\examples{

#================
### Generate data
#================
desc <- cbind(1, scale(1:50), scale(1:50)^2)
nspecies <- 20
commDesc <- communitySimul(X = desc, nsp = nspecies)

#=============
### Formatting
#=============
### Format data
formdata <- as.HMSCdata(Y = commDesc$data$Y, X = desc, interceptX = FALSE, interceptTr=FALSE)

#==============
### Build model
#==============
modelDesc <- hmsc(formdata, niter = 200, nburn = 100, thin = 1, verbose = FALSE)

paramXmcmc <- as.mcmc(modelDesc, parameters = "paramX")

}
\seealso{
\code{\link{mcmc}}, \code{\link{hmsc}}
}
\author{
F. Guillaume Blanchet
}
\keyword{IO}
