% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.hmsc.R
\name{biplot.hmsc}
\alias{biplot.hmsc}
\title{Biplot for HMSC}
\usage{
\method{biplot}{hmsc}(x, Random, choice = c(1, 2), display = c("sites",
  "species"), type, xlim, ylim, xlab, ylab, col = c(1, 2), ...)
}
\arguments{
\item{x}{An object of the class \code{hmsc}.}

\item{Random}{Random effect for which the biplot needs to be drawn. This could be either a character string defining the name of the random effect to consider or a value defining the order of the random effect to consider.}

\item{choice}{Latent variables to show.}

\item{display}{Scores shown. These must some of the alternatives "species" for species scores, and/or "sites" for site scores.}

\item{type}{Type of plot: partial match to text for text labels, points for points, and none for setting frames only. If omitted, text is selected for smaller data sets, and points for larger. Can be of length 2 (e.g. type = c("text", "points")), in which case the first element describes how species scores are handled, and the second how site scores are drawn.}

\item{xlim}{The x limits (min, max) of the plot.}

\item{ylim}{The y limits (min, max) of the plot.}

\item{col}{Colours used for sites and species (in this order). If only one colour is given, it is used for both.}

\item{\dots}{Other parameters for plotting functions.}
}
\value{
An ordination plot for the unconstrained part of the HMSC analysis
}
\description{
Draw a biplot for HMSC analysis carried out on a model with that includes at least one random effect
}
\details{
In the current biplot, the "species" scores are drawn as points and the point position is obtained from \code{paramLatent}. Similarly, the "sites" scores are also point and their location is obtained from \code{latent}.

All the calculations are carried out on the estimation part of the model, the burning part is discarded.

Also, the current function is not yet design to draw biplot using any other random effect than the non-autocorrelated random effect.

Graphical parameters can also be given to biplot: the size of xlabs and ylabs is controlled by cex.
}
\examples{

#==================
### Simulating data
#==================
desc <- matrix(1, nrow = 50, ncol = 1)
random1 <- as.factor(1:50)
random2 <- as.factor(rep(letters[1:2], each = 25))
randEff <- data.frame(rand1 = random1, rand2 = random2)
nspecies <- 10

comm <- communitySimul(X = desc, Random = randEff, nsp = nspecies)

#=================================
### Formatting data and parameters
#=================================
formdata <- as.HMSCdata(Y = comm$data$Y, X = desc, Random = randEff, interceptX = FALSE)

#==============
### Build model
#==============
model <- hmsc(formdata)

#===============================
### Construct correlation matrix
#===============================
biplot(model,Random=1)

}
\references{
Hui, F. K. C., S. Taskinen, S. Pledger, S. D. Foster, and D. I. Warton. 2015. Model-based approaches to unconstrained ordination. \emph{Methods in Ecology and Evolution} \strong{6}, 399–411.

Warton, D. I., F. G. Blanchet, R. B. O’Hara, O. Ovaskainen, S. Taskinen, S. C. Walker, and F. K. C. Hui. 2015. So Many Variables: Joint Modeling in Community Ecology. \emph{Trends in Ecology & Evolution} \strong{30}:766–779.
}
\author{
F. Guillaume Blanchet
}
\keyword{hplot}
