% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.hmsc}
\alias{coef.hmsc}
\title{Extract Model Coefficients for HMSC objects}
\usage{
\method{coef}{hmsc}(object, ...)
}
\arguments{
\item{object}{Object of class hmsc.}

\item{\dots}{other arguments.}
}
\value{
An object of class \code{HMSCparam}.
}
\description{
This function extract the average model coefficients from an \code{HMSC} object
}
\details{
This functions should be used when the MCMC run have converged properly for all parameters.

The output of this function is an object of class \code{HMSCparam}.

For models that include latent variables or autocorrelated latent variables, the latent variables and their associated parameters are also returned.
}
\examples{
#================
### Generate data
#================
desc <- cbind(scale(1:50), scale(1:50)^2)
nspecies <- 5
commDesc <- communitySimul(X = desc, nsp = nspecies)

#=============
### Formatting
#=============
### Format data
formdata <- as.HMSCdata(Y = commDesc$data$Y, X = desc, interceptX = FALSE)

#==============
### Build model
#==============
modelDesc <- hmsc(formdata, niter = 2000, nburn = 1000, thin = 1)

#==============================
### Check for model convergence
#==============================
plot(as.mcmc(modelDesc,parameters = "paramX"))
plot(as.mcmc(modelDesc,parameters = "meansParamX"))
plot(as.mcmc(modelDesc,parameters = "varX"))

#===============================
### Extract average coefficients
#===============================
modelCoef <- coef(modelDesc)

}
\seealso{
\code{\link{hmsc}}

\code{\link{predict.hmsc}}
}
\author{
F. Guillaume Blanchet
}
\keyword{model}
\keyword{regression}
