% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineHmscChains.R
\name{combineHmscChains}
\alias{combineHmscChains}
\title{combineHmscChains}
\usage{
combineHmscChains(hMList, alignPost = TRUE)
}
\description{
Combines different chains of Hmsc model fitted with different arguments in sampleMcmc(..., indChains, ...) call
}
