% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDataParameters.R
\name{computeDataParameters}
\alias{computeDataParameters}
\title{computeDataParameters}
\usage{
computeDataParameters(hM, compactFormat = FALSE)
}
\arguments{
\item{hM}{a fitted \code{Hmsc} model object}

\item{compactFormat}{bool flag whether to get reduced set of parameters - as used in export to Hmsc-HPC}
}
\value{
a list including pre-computed matrix inverses and determinants (for phylogenetic and spatial random effects) needed in MCMC sampling
}
\description{
Computes initial values before the sampling starts
}
