% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMarginalEffects.R
\name{computeMarginalEffects}
\alias{computeMarginalEffects}
\title{Hmsc$computeMarginalEffects}
\usage{
computeMarginalEffects(ngrid = 20, prob = c(0.025, 0.5, 0.975),
  clist = 1:self$nc)
}
\arguments{
\item{clist}{}
}
\value{

}
\description{
Computes ...
}
\examples{

}
\seealso{

}
