% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeR2.R
\name{computeR2}
\alias{computeR2}
\title{Hmsc$computeR2}
\usage{
computeR2(hM, predY)
}
\arguments{
\item{predY}{vector of predictions from the trained model}
}
\value{

}
\description{
Computes explainatory R2 measure based on predicted species abundance matrix
}
\examples{

}
\seealso{

}
