% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeVariancePartitioning.R
\name{computeVariancePartitioning}
\alias{computeVariancePartitioning}
\title{Hmsc$computeVariancePartitioning}
\usage{
computeVariancePartitioning(group, groupnames, start = 1)
}
\arguments{
\item{start}{}
}
\description{
Computes variance partitionint w.r.t. given grouping of fixed efffects and levels of random effects
}
\examples{

}
