% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWAIC.R
\name{computeWAIC}
\alias{computeWAIC}
\title{Hmsc$computeWAIC}
\usage{
computeWAIC(hM)
}
\value{
the scalar WAIC
}
\description{
Computes WAIC (Widely Applicable Information Criterion) of the HMSC model
}
\details{
implemented yet only for probit and normal models
}
\examples{

WAIC = computeWAIC(hM=m)

}
