% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructKnots.R
\name{constructKnots}
\alias{constructKnots}
\title{constructKnots}
\usage{
constructKnots(sData, nKnots = NULL, knotDist = NULL,
  minKnotDist = NULL)
}
\arguments{
\item{sData}{a dataframe containing spatial or temporal coordinates of units of the random level}

\item{minKnotDist}{}
}
\value{
a data frame with knot locations
}
\description{
construct a regular spaced grid with knot locations to be used in spatial Hmsc models with the spatial method set to GPP
}
\details{
to be added
}
\examples{
# Setting a random level with 50 units



}
