% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToCodaObject.r
\name{convertToCodaObject}
\alias{convertToCodaObject}
\title{Hmsc$convertToCodaObject}
\usage{
convertToCodaObject(hM, start = 1, spNamesNumbers = c(TRUE, TRUE),
  covNamesNumbers = c(TRUE, TRUE), trNamesNumbers = c(TRUE, TRUE),
  Beta = TRUE, Gamma = TRUE, V = TRUE, Sigma = TRUE, Rho = TRUE,
  Eta = TRUE, Lambda = TRUE, Alpha = TRUE, Omega = TRUE,
  Psi = TRUE, Delta = TRUE)
}
\arguments{
\item{start}{index of first MCMC sample included}

\item{Delta}{}
}
\value{
A named list that can be analysed with \CRANpkg{coda} functions.
}
\description{
Converts Hmsc posterior structure into a named list of mcmc.list objects
}
\examples{
# Convert recorded posterior samples in m object to coda object
codaObject = convertToCodaObject(TD$m)

# Convert recorded posterior samples, starting from sample 100, in m object to coda object
codaObject = convertToCodaObject(TD$m,start=100)

}
