% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralCombine.R
\name{coralCombine}
\alias{coralCombine}
\title{coralCombine}
\usage{
coralCombine(m, muList.coral, VList.coral)
}
\arguments{
\item{m}{fitted \code{Hmsc}-class object}

\item{muList.coral}{arg2}

\item{VList.coral}{arg3}
}
\value{
list with combined means and covariance matrices
}
\description{
Extracts CORAL-like summary from backbone fitted \code{Hmsc}-class object and combines with CORAL models
}
