% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralGetRareSpeciesPriors.R
\name{coralGetRareSpeciesPriors}
\alias{coralGetRareSpeciesPriors}
\title{coralGetRareSpeciesPriors}
\usage{
coralGetRareSpeciesPriors(
  m,
  spNames.rare,
  TrData.rare = NULL,
  phyloTree = NULL,
  C.common.rare = NULL,
  interceptVarInflation = 5
)
}
\arguments{
\item{m}{fitted \code{Hmsc}-class object}

\item{spNames.rare}{arg2}

\item{TrData.rare}{arg3}

\item{phyloTree}{arg4}

\item{C.common.rare}{arg5}

\item{interceptVarInflation}{arg6}
}
\value{
A list containing means and covariances of CORAL prior
}
\description{
Constructs CORAL priors
}
