% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralGetXDataExt.R
\name{coralGetXDataExt}
\alias{coralGetXDataExt}
\title{coralGetXDataExt}
\usage{
coralGetXDataExt(m, nf = NULL, varProp = NULL)
}
\arguments{
\item{m}{fitted \code{Hmsc}-class object}

\item{nf}{arg2}

\item{varProp}{arg3}
}
\value{
A named list of extended XData data.frame and extended XFormula
}
\description{
Splits Hmsc data into common and rare parts for CORAL analysis
}
