% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralPlotBeta.R
\name{coralPlotBeta}
\alias{coralPlotBeta}
\title{coralPlotBeta}
\usage{
coralPlotBeta(
  mu,
  V,
  phyloTree,
  spNames.common,
  plotColumns = c(1:ncol(mu)),
  quantile.support = c(0.05, 0.95),
  plotTree = 0.3,
  seed = NULL,
  col.common = c("blue", "red"),
  alpha.common = 0.5,
  jitter.common = 0.45,
  cex.common = 0.5,
  pch.common = 16,
  col.rare = c("cyan", "pink"),
  alpha.rare = 0.5,
  jitter.rare = 0.2,
  cex.rare = 0.2,
  pch.rare = 16,
  showCovNames = TRUE
)
}
\arguments{
\item{mu, V}{coral plot arguments, probably}

\item{phyloTree}{phylogenetic tree}

\item{spNames.common}{species names}

\item{plotColumns, quantile.support, plotTree}{arguments to plot}

\item{seed}{seed}

\item{col.common, alpha.common, jitter.common, cex.common, pch.common}{arguments
for common species}

\item{col.rare, alpha.rare, jitter.rare, cex.rare, pch.rare}{arguments for rare
species}

\item{showCovNames}{show covariate names}
}
\value{
list with combined means and covariance matrices
}
\description{
Plots summary of Beta coefficents fitted with CORAL approach
}
