% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralPredict.R
\name{coralPredict}
\alias{coralPredict}
\title{coralPredict}
\usage{
coralPredict(XData, XFormula, mu, V, prob = TRUE)
}
\arguments{
\item{XData}{covariates data.frame}

\item{XFormula}{arg2}

\item{mu}{arg3}

\item{V}{arg4}

\item{prob}{arg5}
}
\value{
Matrix of size ny x ns with CORAL predictions
}
\description{
Makes predictions from CORAL models
}
