% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralPreprocess.R
\name{coralPreprocess}
\alias{coralPreprocess}
\title{coralPreprocess}
\usage{
coralPreprocess(
  Y,
  spNames.common,
  TrData = NULL,
  phyloTree = NULL,
  Taxa = NULL,
  TaxaFormula = NULL
)
}
\arguments{
\item{Y}{arg1}

\item{spNames.common}{arg2}

\item{TrData}{arg3}

\item{phyloTree}{arg4}

\item{Taxa}{arg5}

\item{TaxaFormula}{arg6}
}
\value{
A named list of splitted parts
}
\description{
Splits Hmsc data into common and rare parts for CORAL analysis
}
