% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralSplitToBatches.R
\name{coralSplitToBatches}
\alias{coralSplitToBatches}
\title{coralSplitToBatches}
\usage{
coralSplitToBatches(Y, TrData, C.common.rare, batchN = 1)
}
\arguments{
\item{Y}{fitted \code{Hmsc}-class object}

\item{TrData}{arg2}

\item{C.common.rare}{arg3}

\item{batchN}{arg4}
}
\value{
A list containing three lists with splits of the input arguments
#'
}
\description{
Splits Hmsc data into common and rare parts for CORAL analysis
}
