% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPartition.R
\name{createPartition}
\alias{createPartition}
\title{createPartition}
\usage{
createPartition(hM, nfolds = 10, column = NULL)
}
\arguments{
\item{hM}{a fitted \code{Hmsc} model object}

\item{nfolds}{number of cross-validation folds}

\item{column}{name or index of the column in the \code{studyDesign} matrix, corresponding to the level for which units are splitted into folds}
}
\value{
a vector describing the fold of each sampling unit
}
\description{
Constructs a partition vector given the number of folds and column of study design
}
\examples{
# Create 3 folds for a HMSC object
partition = createPartition(TD$m, nfolds = 3)
}
