% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateModelFit.R
\name{evaluateModelFit}
\alias{evaluateModelFit}
\title{Hmsc$evaluateModelFit}
\usage{
evaluateModelFit(hM, predY)
}
\arguments{
\item{predY}{array of predictions, typically posterior sample}
}
\value{

}
\description{
Computes several measures of model fit
}
\examples{

}
\seealso{

}
