% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatPriorAuto.R, R/flatPriorPhylo.R,
%   R/flatPriorRandom.R, R/flatPriorResidVar.R, R/flatPriorX.R, R/flatPriorXTr.R
\name{flatPriorAuto}
\alias{flatPriorAuto}
\alias{flatPriorPhylo}
\alias{flatPriorRandom}
\alias{flatPriorResidVar}
\alias{flatPriorX}
\alias{flatPriorXTr}
\title{Define flat priors}
\usage{
flatPriorAuto(data, paramAutoDist = NULL, paramAutoWeight = NULL,
  shrinkOverallAuto = NULL, shrinkSpeedAuto = NULL,
  shrinkLocalAuto = NULL, family = "probit")

flatPriorPhylo(paramPhylo = NULL)

flatPriorRandom(shrinkOverall = NULL, shrinkSpeed = NULL,
  shrinkLocal = NULL, family = "probit")

flatPriorResidVar(residVar = NULL)

flatPriorX(varXDf = NULL, varXScaleMat = NULL, meansParamX = NULL,
  varMeansParamX = NULL, nparamX)

flatPriorXTr(varXDf = NULL, varXScaleMat = NULL, paramTr = NULL,
  varTr = NULL, nTr, nparamX)
}
\arguments{
\item{data}{An object of the class \code{HMSCdata}.}

\item{paramAutoDist}{Prior. Matrix with a single column defining the potential values given in \code{paramAuto} can take. The size of \code{priorParamAutoGrid} needs to be equal to the size of \code{priorParamAutoWeight}.}

\item{paramAutoWeight}{Prior. Matrix with a single column defining the importance each value in \code{priorParamAutoDist} may have. The size of \code{priorParamAutoWeight} needs to be equal to the size of \code{priorParamAutoGrid}.}

\item{shrinkOverallAuto}{A vector of length 2 defining the shape and scale hyperparameter of a gamma distribution. The size of the first parameters indicate how the overall shrinkage of the autocorrelated latent variable is handled.}

\item{shrinkSpeedAuto}{A vector of length 2 defining the shape and scale hyperparameter of a gamma distribution. The size of the first parameters indicate how the speed of the shrinkage of the autocorrelated latent variable is handled.}

\item{shrinkLocalAuto}{List. Each level of the list includes a matrix of values describing how much shrinkage should be imposed on each parameter of the autocorrelated latent (unsampled) variables (\code{paramLatentAuto}). There are as many level in the list as there are columns in \code{Auto}. The size of each matrix is the same as the size of \code{paramLatentAuto}.}

\item{family}{Character string defining the type of generalized regression model for which priors need to be defined.}

\item{paramPhylo}{Prior information associated to the parameter of the same name.}

\item{shrinkOverall}{A vector of length 2 defining the shape and scale hyperparameter of a gamma distribution. The size of the first parameters indicate how the overall shrinkage of latent variable is handled.}

\item{shrinkSpeed}{A vector of length 2 defining the shape and scale hyperparameter of a gamma distribution. The size of the first parameters indicate how the speed of the shrinkage of latent variable is handled.}

\item{shrinkLocal}{Numeric. Hyperparameter of a gamma distribution defining the local shrinkage for latent variables.}

\item{residVar}{Prior information associated to the parameter of the same name.}

\item{varXDf}{Numeric. Prior for varX. It is the number of degrees of freedom of the inverse Wishart distribution from which varX is sampled.}

\item{varXScaleMat}{Symmetric covariance matrix. Prior for varX. It is the scale matrix of the inverse Wishart distribution from which varX is sampled.}

\item{meansParamX}{A vector of parameters that defines the prior information associated to parameters of the same name.}

\item{varMeansParamX}{Symmetric covariance matrix. Prior for meansParamX defining how the sampling of the parameters vary.}

\item{nparamX}{Numeric. Number of covariates.}

\item{paramTr}{Matrix of prior defining how descriptors (rows) characterizes traits (columns).}

\item{varTr}{Symmetric covariance matrix. Prior for paramTr defining how the sampling of the parameters vary.}

\item{nTr}{Numeric. Number of traits.}
}
\description{
These functions define flat prior for the different models that can be constructed within the HMSC framework, they are meant to be used internally.
}
\author{
F. Guillaume Blanchet
}
