% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hmsc.R
\name{Hmsc}
\alias{Hmsc}
\title{Hmsc}
\usage{
Hmsc(Y = NULL, XFormula = ~., XData = NULL, X = NULL,
  XScale = TRUE, dfPi = NULL, rL = NULL, Xs = NULL, Xv = NULL,
  TrFormula = NULL, TrData = NULL, Tr = NULL, TrScale = TRUE,
  phyloTree = NULL, C = NULL, distr = "normal", spNames = NULL,
  trNames = NULL, covNames = NULL, levelNames = NULL,
  priors = NULL)
}
\arguments{
\item{Y}{matrix of species occurences or abundances}

\item{XFormula}{formula for linear regression component of HMSC}

\item{XData}{dataframe of measured covariates}

\item{X}{matrix of measured covariates for direct specification}

\item{XScale}{(boolean; default is TRUE) scale values in X matrix?}

\item{dfPi}{matrix of correspondence between sampling units and units on
different levels of latent factors}

\item{rL}{list of HmscRandomLevel objects, specifying the structure and data for random levels}

\item{TrScale}{(boolean; default is TRUE) scale values in the trait (TR) matrix?}

\item{levelNames}{}
}
\value{
Object of Hmsc class
}
\description{
Creates a Hmsc object
}
\examples{

}
\seealso{

}
