% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmscKroneckerRandomLevel.R
\name{HmscKroneckerRandomLevel}
\alias{HmscKroneckerRandomLevel}
\title{Create an \code{Hmsc} kronecker random level structure}
\usage{
HmscKroneckerRandomLevel(
  rLList,
  sepStr = "=",
  etaMethod = "R",
  alphaMethod = "R",
  conserveMemoryFlag = FALSE,
  logDetKstBatchSize = 0,
  cgIterN = NULL
)
}
\arguments{
\item{rLList}{}
}
\value{
a \code{HmscKroneckerRandomLevel}-class object that can be used for \code{Hmsc}-class object construction
}
\description{
Specifies the a kronecker random factor level
}
\details{
Nothing yet.
}
\examples{
# Setting a random level with 50 units
rL = HmscRandomLevel(units=TD$studyDesign$sample)

# Setting a spatial random level
rL = HmscRandomLevel(sData=TD$xycoords)

# Setting a covariate-dependent random level.
rL = HmscRandomLevel(xData=data.frame(x1=rep(1,length(TD$X$x1)),x2=TD$X$x2))

}
\seealso{
[setPriors.Hmsc()]
}
