% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmscRandomLevel.R
\name{HmscRandomLevel}
\alias{HmscRandomLevel}
\title{Create single level of random factors in HMSC}
\usage{
HmscRandomLevel(sData = NULL, distMat = NULL, xData = NULL,
  units = NULL, N = NULL)
}
\arguments{
\item{sData}{a dataframe containing spatial or temporal coordinates of units of the random level}

\item{distMat}{a distance matrix containing the distances between units of the random level}

\item{xData}{a dataframe containing the covariates measured at the units of the random level for covariate-dependent
associations}

\item{units}{a vector, specifying the names of the units of a non-structured level}

\item{N}{number of unique units on this level}
}
\value{
a \code{HmscRandomLevel}-class object that can be used for \code{Hmsc}-class object construction
}
\description{
Specifies the structure for a single level of random factors, including whether the level is assumed to
  be spatial or not, the spatial coordinates and the potential structure of covariate-dependent nonstationarity.
}
\details{
Only one of \code{sData}, \code{distMat}, \code{xData}, \code{units} and \code{N} arguments shall be
  provided (implmentation for \code{sData} and \code{xData} is coming later).

  As a good practice we recommend to specify all available units for a random level, even if some of those are not
  used for training the model.
}
\examples{
# Setting a random level with 50 units
rL = HmscRandomLevel(units=TD$studyDesign$sample)

# Setting a spatial random level
rL = HmscRandomLevel(sData=TD$xycoords)

# Setting a covariate dependent random level.
rL = HmscRandomLevel(xData=data.frame(x1=rep(1,length(TD$X$x1)),x2=TD$X$x2))

}
\seealso{
[setPriors.Hmsc()]
}
