% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmscSimil.R
\name{hmscSimil}
\alias{hmscSimil}
\title{Community similarity induced by environmental similarity}
\usage{
hmscSimil(hmsc, gr = NULL)
}
\arguments{
\item{hmsc}{An object of the class \code{hmsc}.}

\item{gr}{A factor defining groups of samples. If \code{NULL}, all samples are considered as independent groups.}
}
\value{
A symmetrical covariance matrix
}
\description{
Calculates a community similarity resulting from environmental similarity.
}
\details{

}
\examples{

}
\author{
F. Guillaume Blanchet
}
\keyword{datagen}
\keyword{htest}
\keyword{multivariate}
\keyword{regression}
\keyword{univar}
