% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPosteriorFromHPC.R
\name{importPosteriorFromHPC}
\alias{importPosteriorFromHPC}
\title{importPosteriorFromHPC}
\usage{
importPosteriorFromHPC(
  m,
  postList,
  nSamples,
  thin,
  transient,
  alignPost = TRUE
)
}
\arguments{
\item{m}{the \code{Hmsc} model object used for creating export to Hmsc-HPC fitting}

\item{postList}{list of MCMC chains, containing the recorded samples from the MCMC sampling}

\item{nSamples}{the number of recorded MCMC samples per chain, as was used for Hmsc-HPC fitting}

\item{thin}{the MCMC thinning value, as was used for Hmsc-HPC fitting}

\item{transient}{the MCMC transient value, as was used for Hmsc-HPC fitting}

\item{alignPost}{bool flag whether to attempt to align the latent factors and loadings from different chains}
}
\description{
Imports the posterior object after running MCMC using the Hmsc-HPC add-on
}
