% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniParamAuto.R, R/iniParamPhylo.R,
%   R/iniParamRandom.R, R/iniParamResidVar.R, R/iniParamX.R, R/iniParamXTr.R
\name{iniParamAuto}
\alias{iniParamAuto}
\alias{iniParamPhylo}
\alias{iniParamRandom}
\alias{iniParamResidVar}
\alias{iniParamX}
\alias{iniParamXTr}
\title{Initiate parameters}
\usage{
iniParamAuto(data, priors, paramAuto = NULL, latentAuto = NULL,
  paramLatentAuto = NULL, shrinkLocalAuto = NULL,
  paramShrinkGlobalAuto = NULL)

iniParamPhylo(priors, paramPhylo = NULL)

iniParamRandom(data, priors, latent = NULL, paramLatent = NULL,
  shrinkLocal = NULL, paramShrinkGlobal = NULL)

iniParamResidVar(data, residVar = NULL)

iniParamX(data, priors, paramX = NULL, meansParamX = NULL, varX = NULL)

iniParamXTr(data, priors, paramX = NULL, varX = NULL, paramTr = NULL)
}
\arguments{
\item{data}{An object of the class \code{HMSCdata}.}

\item{priors}{An object of the class \code{HMSCprior}.}

\item{paramAuto}{Vector of numerical values defining the autocorrelated patterns for each factor in \code{Auto}. These values can range from 0 to the largest distance between samples in the autocorrelated level.}

\item{latentAuto}{List. Each level of the list includes a matrix of autocorrelated latent (unsampled) variables. There are as many level in the list as there are columns in \code{Auto}. The number of rows in each matrix corresponds to the number of levels in each autocorrelated factor in \code{Auto}.}

\item{paramLatentAuto}{List. Each level of the list includes a matrix of parameters for the autocorrelated latent (unsampled) variables. There are as many level in the list as there are columns in \code{Auto}. Each matrix has as many rows as there are species in \code{Y} and as many columns as there are autocorrelated latent variables in \code{latentAuto}.}

\item{shrinkLocalAuto}{List. Each level of the list includes a matrix of values describing how much shrinkage should be imposed on each parameter of the autocorrelated latent (unsampled) variables (\code{paramLatentAuto}). There are as many level in the list as there are columns in \code{Auto}. The size of each matrix is the same as the size of \code{paramLatentAuto}.}

\item{paramShrinkGlobalAuto}{List. Each level of the list includes a vector of values describing how much shrinkage should be imposed on each autocorrelated latent variables (\code{latent}) as a whole. There are as many level in the list as there are columns in \code{Auto}. The number of elements in each vector equals the number of variables in \code{latentAuto}.}

\item{paramPhylo}{Numeric. Defines the importance of phylogeny in structuring the community.}

\item{latent}{List. Each level of the list includes a matrix of latent (unsampled) variables. There are as many level in the list as there are columns in \code{Random}. The number of rows in each matrix corresponds to the number of levels in each factor in \code{Random}.}

\item{paramLatent}{List. Each level of the list includes a matrix of parameters for the latent (unsampled) variables. There are as many level in the list as there are columns in \code{Random}. Each matrix has as many rows as there are species in \code{Y} and as many columns as there are latent variables in \code{latent}.}

\item{shrinkLocal}{List. Each level of the list includes a matrix of values describing how much shrinkage should be imposed on each parameter of the latent (unsampled) variables (\code{paramLatent}). There are as many level in the list as there are columns in \code{Random}. The size of each matrix is the same as the size of \code{paramLatent}.}

\item{paramShrinkGlobal}{List. Each level of the list includes a vector of values describing how much shrinkage should be imposed on each latent variables (\code{latent}) as a whole. There are as many level in the list as there are columns in \code{Random}. The number of elements in each vector equals the number of variables in \code{latent}.}

\item{residVar}{Vector of parameters with as many entries as there are species. Each values in this vector describes how much variation is not explained by the model.}

\item{paramX}{A matrix or data.frame of model parameters defining how species (rows) are characterized by the descriptors (columns).}

\item{meansParamX}{A vector of parameters defining how an average species reacts to each descriptor in X.}

\item{varX}{Symmetric covariance matrix. Each dimension of this matrix should be equal to the number of explanatory variables.}

\item{paramTr}{Matrix of model parameters defining how descriptors (rows) characterizes traits (columns).}
}
\description{
These functions are used to initiate the different sets of parameters used to construct different type of models within the HMSC framework, they are meant to be used internally.
}
\author{
F. Guillaume Blanchet
}
