% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniVerbose.R
\name{iniVerbose}
\alias{iniVerbose}
\title{Define what the function should print on screen}
\usage{
iniVerbose(verbose, niter)
}
\arguments{
\item{verbose}{a logical or numerical value}

\item{niter}{Number of of iterations carried out in the analysis}
}
\description{
Defines how the \code{verbose} arguments should be handled. This is meant to be an internal function.
}
\details{
If \code{verbose} is set to  \code{TRUE}, the number of iterations completed is printed on screen after every \code{niter}/5 iterations.
If \code{verbose} is a numerical value, a multiple of this value is printed on screen. 
If the number of iterations is smaller than 5, nothing is printed on screen.
}
\author{
F. Guillaume Blanchet
}
\keyword{print}
