% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniYlatent.R
\name{iniYlatent}
\alias{iniYlatent}
\title{Initiate a latent Y}
\usage{
iniYlatent(data, param, family)
}
\arguments{
\item{data}{An object of the class \code{HMSCdata}.}

\item{param}{An object of the class \code{HMSCparam}.}

\item{family}{A character string defining how the latent response variables needs to be initiated.}
}
\value{
A matrix or a list of objects that are needed to construct the model of interest.
}
\description{
Initiate a latent set of response variables for the different model types
}
\details{
\code{HMSCdata} and \code{HMSCparam}, the proper set of starting parameters are generated.
Since the function is meant to be used by expert users, there are no predefined setting for any of the argument of the function.
}
\author{
F. Guillaume Blanchet
}
\keyword{datagen}

