% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jposterior.R
\name{jposterior}
\alias{jposterior}
\title{Full joint posterior distribution}
\usage{
jposterior(hmsc)
}
\arguments{
\item{hmsc}{An object of the class \code{hmsc}}
}
\value{
A three dimensional array with as first dimension the number of samples in the community matrix \code{Y}, the second dimension the number of species in the community matrix \code{Y} and as third dimension the number of iterations after the burning phase when constructing the model.
}
\description{
Calculate the full joint posterior distribution of the model for each iterations
}
\examples{

#================
### Generate data
#================
desc <- cbind(scale(1:50), scale(1:50)^2)
nspecies <- 20
commDesc <- communitySimul(X = desc, nsp = nspecies)

#=============
### Formatting
#=============
### Format data
formdata <- as.HMSCdata(Y = commDesc$data$Y, X = desc, interceptX = FALSE, interceptTr = FALSE)

#==============
### Build model
#==============
modelDesc <- hmsc(formdata, niter = 2000, nburn = 1000, thin = 1, verbose = 100)

#==============================================
### Calculate full joint posterior distribution
#==============================================
jpost <- jposterior(modelDesc)

}
\author{
F. Guillaume Blanchet
}
\keyword{multivariate,}
\keyword{regression}
\keyword{univar,}
