% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcmcPoissonX}
\alias{mcmcPoissonX}
\title{Markov chain Monte Carlo algorithms.}
\usage{
mcmcPoissonX(Y, Ylatent, X, paramX, meansParamX, precX, residVar,
  priorMeansParamX, priorVarXScaleMat, priorVarXDf, priorResidVarScale,
  priorResidVarShape, nsp, nsite, nparamX, niter, nburn, thin, verbose)
}
\arguments{
\item{Y}{Raw site by species community matrix.}

\item{Ylatent}{Model site by species community matrix after the link function is applied.}

\item{X}{Site by covariates matrix.}

\item{paramX}{Species by covariates matrix of regression parameters.}

\item{meansParamX}{Matrix with a single column and as many rows as covariates. This matrix defines how an average species react to the covariates.}

\item{precX}{Square (precision) matrix with as many rows (and columns) as there are covariates. The inverse of this matrix is a variance matrix which explains how a typical species vary in its response to one (diagonal) or a pair (off-diagonal) of covariates.}

\item{residVar}{Vector of parameters with as many entries as there are species. Each values in this vector describes how much variation is not explained by the model.}

\item{priorMeansParamX}{Prior. Matrix with a single column and as many rows as covariates. It describes the prior values associated to meansParamX.}

\item{priorVarXScaleMat}{Square matrix with as many rows (and columns) as there are covariates. Prior. Scale matrix used to sample \code{precX} from an inverse Wishart distribution.}

\item{priorVarXDf}{Numeric. Prior. Number of degrees of freedoms used to sample \code{precX} from an inverse Wishart distribution.}

\item{priorResidVarScale}{Numeric. Prior. Scale parameter of a gamma distribution from which \code{residVar} is sampled.}

\item{priorResidVarShape}{Numeric. Prior. Shape parameter of a gamma distribution from which \code{residVar} is sampled.}

\item{nsp}{Numeric. Number of species in \code{Y}.}

\item{nsite}{Numeric. Number of site sampled.}

\item{nparamX}{Numeric. Number of covariates.}

\item{niter}{Numeric. Number of iterations performed.}

\item{nburn}{Numeric. Of the number of iterations performed, number of discarded iterations. Burning phase of the model estimations.}

\item{thin}{Numeric. Thinning. Iterations saved every time a multiple of the given value is an integer.}

\item{verbose}{Logical or numeric. If \code{TRUE}, the number of iterations are printed on the screen four times. If \code{FALSE}, nothing is printed on the screen. If a positive integer is given, the number of iterations is printed on the screen every time a multiple of the given value is an integer.}

\item{Tr}{Trait by species matrix.}

\item{Phylo}{Square correlation (with values ranging from -1 to 1) matrix with as many rows (and columns) as there are species in \code{Y}.}

\item{iPhylo}{Square correlation matrix with as many rows (and columns) as there are species in \code{Y}. Inverse of \code{Phylo}.}

\item{Auto}{List with as many levels as there are autocorrelated factors in the analysis. Each level of the list contains a set of coordinates associated to each level of each factor.}

\item{RandomAuto}{A data.frame with all the autocorrelated factors (organized as columns) in the analysis. The number of columns in \code{RandomAuto} equals the number of levels in the \code{Auto}.}

\item{Random}{A data.frame with all the non-autocorrelated factors (organized as columns) in the analysis.}

\item{paramTr}{Covariate by traits matrix of regression parameters.}

\item{paramPhylo}{Numeric. Single parameter describing the importance of phylogeny in the analysis.}

\item{latent}{List. Each level of the list includes a matrix of latent (unsampled) variables. There are as many level in the list as there are columns in \code{Random}. The number of rows in each matrix corresponds to the number of levels in each factor in \code{Random}.}

\item{paramLatent}{List. Each level of the list includes a matrix of parameters for the latent (unsampled) variables. There are as many level in the list as there are columns in \code{Random}. Each matrix has as many rows as there are species in \code{Y} and as many columns as there are latent variables in \code{latent}.}

\item{shrinkLocal}{List. Each level of the list includes a matrix of values describing how much shrinkage should be imposed on each parameter of the latent (unsampled) variables (\code{paramLatent}). There are as many level in the list as there are columns in \code{Random}. The size of each matrix is the same as the size of \code{paramLatent}.}

\item{paramShrinkGlobal}{List. Each level of the list includes a vector of values describing how much shrinkage should be imposed on each latent variables (\code{latent}) as a whole. There are as many level in the list as there are columns in \code{Random}. The number of elements in each vector equals the number of variables in \code{latent}.}

\item{paramAuto}{List.  Each level of the list includes a vector of parameters, one for each autocorrelated latent variables in \code{latentAuto}.  There are as many level in the list as there are columns in \code{Auto}.}

\item{latentAuto}{List. Each level of the list includes a matrix of autocorrelated latent (unsampled) variables. There are as many level in the list as there are columns in \code{Auto}. The number of rows in each matrix corresponds to the number of levels in each autocorrelated factor in \code{Auto}.}

\item{paramLatentAuto}{List. Each level of the list includes a matrix of parameters for the autocorrelated latent (unsampled) variables. There are as many level in the list as there are columns in \code{Auto}. Each matrix has as many rows as there are species in \code{Y} and as many columns as there are autocorrelated latent variables in \code{latentAuto}.}

\item{shrinkLocalAuto}{List. Each level of the list includes a matrix of values describing how much shrinkage should be imposed on each parameter of the autocorrelated latent (unsampled) variables (\code{paramLatentAuto}). There are as many level in the list as there are columns in \code{Auto}. The size of each matrix is the same as the size of \code{paramLatentAuto}.}

\item{paramShrinkGlobalAuto}{List. Each level of the list includes a vector of values describing how much shrinkage should be imposed on each autocorrelated latent variables (\code{latent}) as a whole. There are as many level in the list as there are columns in \code{Auto}. The number of elements in each vector equals the number of variables in \code{latentAuto}.}

\item{priorVarMeansParamX}{square matrix of parameters defining how \code{meansParamX} varies.}

\item{priorParamTr}{Prior. Matrix of priors defining how descriptors (rows) characterizes traits (columns).}

\item{priorVarTr}{Prior. Symmetric covariance matrix of prior. Each dimension of this matrix is be equal to the number of traits.}

\item{priorParamPhyloWeight}{Prior. Matrix with a single column defining the importance each potential phylogenetic weight (given in \code{priorParamPhyloGrid}) may have. The size of \code{priorParamPhyloWeight} needs to be equal to the size of \code{priorParamPhyloGrid}.}

\item{priorParamPhyloGrid}{Prior. Matrix with a single column defining the potential values \code{paramPhylo} can take. The size of \code{priorParamPhyloGrid} needs to be equal to the size of \code{priorParamPhyloWeight}.}

\item{priorParamAutoWeight}{Prior. Matrix with a single column defining the importance each value in \code{priorParamAutoDist} may have. The size of \code{priorParamAutoWeight} needs to be equal to the size of \code{priorParamAutoGrid}.}

\item{priorParamAutoDist}{Prior. Matrix with a single column defining the potential values given in \code{paramAuto} can take. The size of \code{priorParamAutoGrid} needs to be equal to the size of \code{priorParamAutoWeight}.}

\item{priorShrinkLocal}{Numeric. Hyperparameter of a gamma distribution defining the local shrinkage for latent variables.}

\item{priorShrinkOverallShape}{Numeric. Shape of a gamma distribution.}

\item{priorShrinkOverallScale}{Numeric. Scale of a gamma distribution.}

\item{priorShrinkSpeedShape}{Numeric. Shape of a gamma distribution.}

\item{priorShrinkSpeedScale}{Numeric. Scale of a gamma distribution.}

\item{adapt}{Vector of two parameters defining whether the number of latent and autocorrelated latent variables should be modified. The function used is : 1/exp(adapt[1]+(adapt[2]*niter)).}

\item{redund}{Vector of two parameters defining (1) the proportion of redundant elements within latent and autocorrelated latent variables and (2) the amount of error that can be accounted for given the proportion of redundancy.}

\item{nAuto}{Numeric. Number of autocorrelated random effect.}

\item{nAutoLev}{Vector defining the number of levels for each autocorrelated random effect. The length of this vector equals \code{nAuto}.}

\item{nLatentAuto}{Vector defining the number of autocorrelated latent variables for each autocorrelated random effect. The length of this vector equals \code{nAuto}.}

\item{nRandom}{Numeric. Number of non-autocorrelated random effect.}

\item{nRandomLev}{Vector defining the number of levels for each non-autocorrelated random effect. The length of this vector equals \code{nRandom}.}

\item{nLatent}{Vector defining the number of non-autocorrelated latent variables for each non-autocorrelated random effect. The length of this vector equals \code{nRandom}.}

\item{nTr}{Numeric. Number of traits.}

\item{nparamPhylo}{Numeric. Number of values in \code{priorParamPhyloGrid}.}

\item{npriorParamAuto}{Numeric. Number of values in \code{priorParamAutoGrid}.}
}
\description{
These functions are meant to be used internally or by advanced users as they generate raw (somewhat unnamed output)
}

