% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mtest.R
\name{Mtest}
\alias{Mtest}
\title{Matrix Method Tester for Posterior Sampling}
\usage{
Mtest(
  hM,
  initPar = NULL,
  dataParList = NULL,
  updater = list(),
  fromPrior = FALSE
)
}
\arguments{
\item{hM}{a fitted \code{Hmsc} model object}

\item{initPar}{a named list of parameter values used for initiation
of MCMC states}

\item{dataParList}{a named list with pre-computed \code{Qg},
\code{iQg}, \code{RQg}, \code{detQg}, \code{rLPar} parameters}

\item{updater}{a named list, specifying which conditional updaters
should be ommitted}

\item{fromPrior}{whether prior (TRUE) or posterior (FALSE) is to be
sampled}
}
\value{
Sampled list of updated parameters.
}
\description{
Tests the use of \CRANpkg{Matrix} for sampling the
    posterior with block-conditional Gibbs MCMC sampler. The
    function is based on \code{\link{sampleMcmc}}, but only calls
    once the specified samplers which were modified to use
    \CRANpkg{Matrix} functions instead of \pkg{base} and
    \pkg{stats} functions. The results must remain unchanged after
    moving to \CRANpkg{Matrix} and we see if they are any faster.
}
\examples{
data(TD)
set.seed(4711)
system.time(out <- Mtest(TD$m))
out
rm(.Random.seed)

}
\seealso{
\code{\link{sampleMcmc}}, \code{\link{Hmsc}}
}
