% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAuto.R, R/namePhylo.R, R/nameRandom.R,
%   R/nameResidVar.R, R/nameResult.R, R/nameX.R, R/nameXTr.R
\name{nameAuto}
\alias{nameAuto}
\alias{namePhylo}
\alias{nameRandom}
\alias{nameResidVar}
\alias{nameResult}
\alias{nameX}
\alias{nameXTr}
\title{Name result}
\usage{
nameAuto(data, result, niter, nburn, thin, listLev = 1)

namePhylo(data, result, niter, nburn, thin, listLev = 4)

nameRandom(data, result, niter, nburn, thin, listLev = 1)

nameResidVar(data, result, niter, nburn, thin, listLev = 4)

nameResult(data, priors, result, niter, nburn, thin)

nameX(data, result, niter, nburn, thin)

nameXTr(data, result, niter, nburn, thin)
}
\arguments{
\item{data}{An object of the class \code{HMSCdata}.}

\item{result}{An hmsc object without names on any part of the object.}

\item{niter}{An integer defining the number of iterations used to perform the analysis.}

\item{nburn}{An integer defining the number of burning iterations used to perform the analysis.}

\item{thin}{An integer defining thinning in the analysis.}

\item{listLev}{A numeric value defining on which level of the list names should be added.}

\item{priors}{An object of the class \code{HMSCprior}.}
}
\value{
The result object with name for all dimensions and parts of lists
}
\description{
Name all parts of the result object. This is meant to be an internal function.
}
\author{
F. Guillaume Blanchet
}
\keyword{utilities}
