% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthProcrustesRotMat.R
\name{orthProcrustesRotMat}
\alias{orthProcrustesRotMat}
\title{Orthogonal procrustean rotation matrix}
\usage{
orthProcrustesRotMat(X, Z)
}
\arguments{
\item{X}{input matrix}

\item{Z}{target matrix}
}
\value{
the rotation matrix
}
\description{
Orthogonal procrustean rotation matrix
}

