% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGamma.R
\name{plotGamma}
\alias{plotGamma}
\title{plotBeta}
\usage{
plotGamma(post, param = "Gamma", trOrder = "Original", trVector = NULL,
  covOrder = "Original", covVector = NULL, trNamesNumbers = c(T, T),
  covNamesNumbers = c(T, T), supportLevel = 0.9, cex = c(0.8, 0.8, 0.8))
}
\arguments{
\item{cex}{}
}
\description{
Plots betas
}
\examples{

}
