% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGradient.R
\name{plotGradient}
\alias{plotGradient}
\title{Hmsc$plotGradient}
\usage{
plotGradient(hM, Gradient, pred, measure, index = 1, prob = c(0.025,
  0.5, 0.975))
}
\arguments{
\item{prob}{}
}
\value{

}
\description{
Plots ...
}
\examples{

}
\seealso{

}
