% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarginalEffects.R
\name{plotMarginalEffects}
\alias{plotMarginalEffects}
\title{Hmsc$plotMarginalEffects}
\usage{
plotMarginalEffects(pred, covariate, measure, index = 1)
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Plots ...
}
\examples{

}
\seealso{

}
