% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariancePartitioning.R
\name{plotVariancePartitioning}
\alias{plotVariancePartitioning}
\title{plotVariancePartitioning}
\usage{
plotVariancePartitioning(hM, VP, ...)
}
\arguments{
\item{hM}{A fitted HMSC model}

\item{VP}{A HMSC variance partitioning produced by `HMSC::computeVariancePartitioning`}

\item{...}{other parameters passed to the barplot function}
}
\value{

}
\description{
Plots the results of variance partitioning of a HMSC model produced by `HMSC::computeVariancePartitioning` as a barplot
}
\examples{

}
\seealso{

}
