% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolMcmcChains.R
\name{poolMcmcChains}
\alias{poolMcmcChains}
\title{poolMcmcChains}
\usage{
poolMcmcChains(postList, chainIndex = 1:length(postList), start = 1, thin = 1)
}
\arguments{
\item{postList}{list of posterior chains}

\item{chainIndex}{index of chains to be included}

\item{start}{index of first MCMC sample included}

\item{thin}{thinning between included MCMC samples}
}
\value{
a list with combined MCMC samples
}
\description{
Combines a list of single or several MCMC chains into a single chain
}
\examples{
# Combine the posteriors from all chains in a Hmsc object
postList = TD$m$postList
pooledPost = poolMcmcChains(postList)

}
