% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictConditional.R
\name{predictConditional}
\alias{predictConditional}
\title{predictConditional}
\usage{
predictConditional(Yc, post = poolMcmcChains(self$postList), mcmcStep = 1,
  XData = NULL, X = NULL, dfPiNew = self$dfPi, rL = self$rL,
  expected = FALSE, predictEtaMean = FALSE)
}
\arguments{
\item{expected}{(boolean; default is FALSE)}

\item{predictEtaMean}{(boolean; default is FALSE)}
}
\value{

}
\description{
Make conditions predictions given HMSC model object
}
\examples{

}
\seealso{

}
