% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictLatentFactor.R
\name{predictLatentFactor}
\alias{predictLatentFactor}
\title{predictLatentFactor}
\usage{
predictLatentFactor(unitsPred, units, postEta, postAlpha, rL,
  predictMean = FALSE)
}
\arguments{
\item{predictMean}{(boolean; default is FALSE)}
}
\description{
Predicts
}
\examples{

}
