% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareGradient.R
\name{prepareGradient}
\alias{prepareGradient}
\title{prepareGradient}
\usage{
prepareGradient(hM, XDataNew, sDataNew)
}
\arguments{
\item{hM}{a fitted \code{Hmsc} model object}

\item{XDataNew}{a dataframe of the new XData}

\item{sDataNew}{a list of the new sData}
}
\value{
a named list with members \code{XDataNew}, \code{studyDesignNew} and \code{rLNew}
}
\description{
prepares a user-made environmental and/or spatial gradient to be used for prediction
}
\details{
The dataframe \code{XDataNew} is the for output as for input. The main purpose of this function is to prepare
the study design and random levels so that predictions can be made with the \code{\link{predict}} function.
Note that the difference between \code{constructGradient} and \code{prepareGradient} is that while
\code{prepareGradient} takes as input the new environmental and spatial data, \code{constructGradient}
generates those data to represent a new environmental gradient.
}
\seealso{
\code{\link{constructGradient}}, \code{\link{predict}}
}
