% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.HmscRandomLevel.R
\name{print.HmscRandomLevel}
\alias{print.HmscRandomLevel}
\title{print.HmscRandomLevel}
\usage{
\method{print}{HmscRandomLevel}(rL)
}
\arguments{
\item{rL}{HmscRandomLevel object}
}
\description{
print method for HmscRandomLevel
}
\examples{

}
