% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Sample from a multivariate Normal distribution.}
\usage{
rmvnorm(n, Mean, Var)
}
\arguments{
\item{n}{Numeric. Number of samples.}

\item{Mean}{A vector giving the means of the variables sampled.}

\item{Var}{A positive-definite symmetric matrix specifying the covariance matrix of the variables.}
}
\description{
Produces one or more samples from the specified multivariate normal distribution. Unlike the function \code{\link{mvrnorm}[MASS]}, this function is very simple and does not perform the many checks carried out in the \code{\link{mvrnorm}[MASS]}. For this reason, it is faster but it must be used carefully as it may lead to some unexpected error. This function is only meant to be used internally.
}

