% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePrior.R
\name{samplePrior}
\alias{samplePrior}
\title{samplePrior}
\usage{
samplePrior(hM, dataParList = NULL)
}
\arguments{
\item{hM}{a fitted \code{Hmsc} model object}

\item{dataParList}{list of data parameters (see \code{\link{computeDataParameters}})}
}
\value{
A named list containing the Hmsc model parameters
}
\description{
Samples the parameter vector from prior
}
