% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleYlatentBinomialLogit}
\alias{sampleYlatentBinomialLogit}
\title{Sample the model response matrix}
\usage{
sampleYlatentBinomialLogit(Y, Ylatent, EstModel, ncount, nsp, nsite)
}
\arguments{
\item{Y}{The species community matrix to be modelled.}

\item{Ylatent}{Model site by species community matrix after the link function is applied.}

\item{EstModel}{Estimated model for the site by species community matrix.}

\item{ncount}{Numeric. Maximum number of individual for a species.}

\item{nsp}{Numeric. Number of species in \code{Y}.}

\item{nsite}{Numeric. Number of site sampled.}
}
\description{
Sample the model response matrix after the Poisson link function was applied. This function is meant to be used internally.
}
