% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleYlatentProbit}
\alias{sampleYlatentProbit}
\title{Sample the model response matrix}
\usage{
sampleYlatentProbit(Y0Loc, Y1Loc, Ylatent, EstModel, residVar, nsp, nsite)
}
\arguments{
\item{Y0Loc}{A vector defining the locations of all 0s in the community matrix (\code{Y}).}

\item{Y1Loc}{A vector defining the locations of all 1s in the community matrix (\code{Y}).}

\item{Ylatent}{Model site by species community matrix after the link function is applied.}

\item{EstModel}{Estimated model for the site by species community matrix.}

\item{residVar}{Vector of parameters with as many entries as there are species. Each values in this vector describes how much variation is not explained by the model. This vector should only contains 1s, but it was included in this function to deal with potential situations that may arise (in other words, in case I forgot a special case).}

\item{nsp}{Numeric. Number of species in \code{Y}.}

\item{nsite}{Numeric. Number of site sampled.}
}
\description{
Sample the model response matrix after the Probit link function was applied. This function is meant to be used internally.
}

