% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPriors.HmscRandomLevel.R
\name{setPriors.HmscRandomLevel}
\alias{setPriors.HmscRandomLevel}
\title{setPriors.HmscRandomLevel}
\usage{
\method{setPriors}{HmscRandomLevel}(rL, nu = NULL, a1 = NULL,
  a2 = NULL, b1 = NULL, b2 = NULL, alphapw = NULL, nfMax = NULL,
  nfMin = NULL, setDefault = FALSE, ...)
}
\arguments{
\item{nu}{}
}
\value{
Modified HmscRandomLevel object
}
\description{
Sets or resets priors to the Hmsc object
}
