% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPriors.HmscSpatialRandomLevel.R
\name{setPriors.HmscSpatialRandomLevel}
\alias{setPriors.HmscSpatialRandomLevel}
\title{setPriors.HmscSpatialRandomLevel}
\usage{
\method{setPriors}{HmscSpatialRandomLevel}(
  rL,
  nu = NULL,
  a1 = NULL,
  a2 = NULL,
  b1 = NULL,
  b2 = NULL,
  nfMax = NULL,
  nfMin = NULL,
  alphapw = NULL,
  setDefault = FALSE
)
}
\arguments{
\item{rL}{a fitted \code{HmscSpatialRandomLevel} model object}

\item{nu, a1, b1, a2, b2}{parameters of the multiplicative gamma process shrinking prior}

\item{nfMax}{maximum number of latent factors to be sampled}

\item{nfMin}{minimum number of latent factors to be sampled}

\item{alphapw}{discrete grid prior for spatial scale parameter}

\item{setDefault}{logical indicating whether default priors should be used}

\item{...}{other arguments (ignored)}
}
\value{
Modified HmscSpatialRandomLevel object
}
\description{
Sets or resets priors to the Hmsc object
}
