% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Community similarity induced by environmental similarity}
\usage{
similarity(hmsc, gr = NULL)
}
\arguments{
\item{hmsc}{An object of the class \code{hmsc}.}

\item{gr}{A factor defining groups of samples. If \code{NULL}, all samples are considered as independent groups.}
}
\value{
A symmetrical covariance matrix between all pairs of samples
}
\description{
Calculates a community similarity resulting from environmental similarity.
}
\examples{
#================
### Generate data
#================
desc <- cbind(scale(1:50), scale(1:50)^2)
nspecies <- 20
commDesc <- communitySimul(X = desc, nsp = nspecies)

#=============
### Formatting
#=============
### Format data
formdata <- as.HMSCdata(Y = commDesc$data$Y, X = desc, interceptX = FALSE, interceptTr = FALSE)

#==============
### Build model
#==============
modelDesc <- hmsc(formdata, niter = 2000, nburn = 1000, thin = 1, verbose = 100)

#=======================
### Calculate similarity
#=======================
similarity <- similarity(modelDesc)

}
\author{
F. Guillaume Blanchet
}
\keyword{datagen}
\keyword{htest}
\keyword{multivariate}
\keyword{regression}
\keyword{univar}
