% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignPosterior.R
\name{alignPosterior}
\alias{alignPosterior}
\title{alignPosterior}
\usage{
alignPosterior(hM)
}
\arguments{
\item{hM}{a fitted \code{Hmsc} model object}
}
\value{
an \code{Hmsc} model object that is identical to the input except the posterior being aligned
}
\description{
Aligns posterior in terms of variables susceptible to label switching
}
\examples{
# Align the posterior for a previously fitted HMSC model
m = alignPosterior(TD$m)

}
