% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.mcmc.hmsc.R
\name{as.mcmc.hmsc}
\alias{as.mcmc.hmsc}
\title{Convert MCMC parameters estimation to \code{\link{mcmc}} object}
\usage{
\method{as.mcmc}{hmsc}(x, parameters = "paramX", burning = FALSE, ...)
}
\arguments{
\item{x}{An object of the class \code{hmsc}.}

\item{parameters}{A character string defining the parameters for which the convergions needs to be carried out.}

\item{burning}{Logical. Whether the burning iterations should be include (\code{TRUE}) or not (\code{FALSE}).}

\item{\dots}{Addtional arguments passed to \code{\link[coda]{mcmc}}.}
}
\value{
An object of class \code{\link{mcmc}}.
}
\description{
Use the Markov Chain Monte Carlo results obtained from \code{\link{hmsc}} and convert them to an \code{\link{mcmc}} object.
}
\details{
The parameters estimated for the latent variables (\code{paramLatent}) cannot be directly converted to an object of class \code{mcmc} because the procedure implemented in \code{\link{hmsc}} allows for the number of latent variables, and thus of parameters associated to these latent variables, to change. For this reason, if \code{paramLatent} is used, it is the matrix defining the correlations among species calculated from  \code{paramLatent} for which an \code{\link{mcmc}} object is constructed. 

The \code{\link{mcmc}} object associated to \code{paramLatent} includes information only for the correlations among single pairs of species, that is, only the lower triangle of the correlation matrix is considered.

When there are multiple random effects estimated for the model, associated to each random effect there is a set of estimated latent parameters (\code{paramLatent}). For this reason, when the Markov Chain Monte Carlo results for \code{paramLatent} is converted with \code{as.mcmc}, parameters associated to each set of latent variables in \code{paramLatent} becomes an code{\link{mcmc}} object in a list. It is important to note that this "list" is \emph{NOT} an \code{mcmc.list}.
}
\examples{

#================
### Generate data
#================
desc <- cbind(1, scale(1:50), scale(1:50)^2)
nspecies <- 20
commDesc <- communitySimul(X = desc, nsp = nspecies)

#=============
### Formatting
#=============
### Format data
formdata <- as.HMSCdata(Y = commDesc$data$Y, X = desc, interceptX = FALSE, 
						   interceptTr=FALSE)

#==============
### Build model
#==============
modelDesc <- hmsc(formdata, niter = 200, nburn = 100, thin = 1, verbose = FALSE)

paramXmcmc <- as.mcmc(modelDesc, parameters = "paramX")

}
\seealso{
\code{\link{mcmc}}, \code{\link{hmsc}}
}
\author{
F. Guillaume Blanchet
}
\keyword{IO}
