% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biPlot.R
\name{biPlot}
\alias{biPlot}
\title{biPlot}
\usage{
biPlot(hM, etaPost, lambdaPost, factors = c(1, 2), colVar = NULL, ...)
}
\arguments{
\item{etaPost}{Posterior distribution of site loadings Eta}

\item{lambdaPost}{Posterior distribution of species loadings Lambda}

\item{factors}{Indices of the two factors to be plotted}

\item{colvar}{The environmental covariate from XData according to which the sites are to be coloured}
}
\description{
Constructs a ordination biplot based on the fitted model
}
\examples{
\dontrun{
biPlot(m, etaPost = getPostEstimate(m, "Eta"), lambdaPost=getPostEstimate(m, "Lambda"), factors=c(1,2))
}

}
