% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAssociations.R
\name{computeAssociations}
\alias{computeAssociations}
\title{computeAssociations}
\usage{
computeAssociations(hM, start = 1)
}
\arguments{
\item{start}{index of first MCMC sample included}
}
\value{
list of association matrices (\eqn{\omega}) corresponding to each random level in the model
}
\description{
Computes the species association matrices
}
\examples{

}
\seealso{

}
