% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAssociations.R
\name{computeAssociations}
\alias{computeAssociations}
\title{computeAssociations}
\usage{
computeAssociations(hM, start = 1, thin = 1)
}
\arguments{
\item{hM}{a fitted \code{Hmsc} model object}

\item{start}{index of first MCMC sample included}

\item{thin}{thinning interval of posterior distribution}
}
\value{
list of association matrices (\eqn{\omega}) corresponding to each random level in the model
}
\description{
Computes the species association matrices associated with each random level
}
\examples{
# Compute the associations (residual correlations) between species from a HMSC model
assoc = computeAssociations(TD$m)

}
