% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDataParameters.R
\name{computeDataParameters}
\alias{computeDataParameters}
\title{computeInitialParameters}
\usage{
computeDataParameters(hM)
}
\arguments{
\item{initPar}{initial parameters value}
}
\description{
Computes the initial values before the sampling starts
}
