% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInitialParameters.R
\name{computeInitialParameters}
\alias{computeInitialParameters}
\title{computeInitialParameters}
\usage{
computeInitialParameters(initPar)
}
\arguments{
\item{initPar}{initial parameters value}
}
\description{
Computes the initial values before the sampling starts
}
\examples{

}
