% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInitialParameters.R
\name{computeInitialParameters}
\alias{computeInitialParameters}
\title{computeInitialParameters}
\usage{
computeInitialParameters(hM, initPar, computeZ = TRUE)
}
\arguments{
\item{hM}{a \code{Hmsc} model object}

\item{initPar}{a list of initial parameter values}

\item{computeZ}{whether to compute the latent predictor Z. Disable only for the purpose of Hmsc-HPC initialization.}
}
\value{
a list of Hmsc model parameters
}
\description{
Computes initial parameter values before the sampling starts
}
