% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInitialParameters.R
\name{computeInitialParameters}
\alias{computeInitialParameters}
\title{computeInitialParameters}
\usage{
computeInitialParameters(hM, initPar)
}
\arguments{
\item{hM}{a fitted \code{Hmsc} model object}

\item{initPar}{initial parameter values}
}
\value{
a list of Hmsc model parameters
}
\description{
Computes initial parameter values before the sampling starts
}
