% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePredictedValues.R
\name{computePredictedValues}
\alias{computePredictedValues}
\title{Hmsc$computePredictedValues}
\usage{
computePredictedValues(nfolds = NULL, column = NULL, partition = NULL,
  start = 1, expected = TRUE)
}
\arguments{
\item{nfolds}{number of cross-validation folds}

\item{start}{index of first MCMC sample included}
}
\description{
Computes predicted values from the trained model
}
\examples{

}
