% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeVariancePartitioning.R
\name{computeVariancePartitioning}
\alias{computeVariancePartitioning}
\title{Hmsc$computeVariancePartitioning}
\usage{
computeVariancePartitioning(hM, group, groupnames, start = 1)
}
\arguments{
\item{group}{vector of numeric values corresponding to group identifiers in groupnames}

\item{groupnames}{vector of names for each random and fixed effect}

\item{start}{index of first MCMC sample included}
}
\value{

}
\description{
Computes variance partitions with respect to given grouping of fixed efffects and levels of random effects
}
\examples{

}
\seealso{

}
