% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWAIC.R
\name{computeWAIC}
\alias{computeWAIC}
\title{Hmsc$computeWAIC}
\usage{
computeWAIC(hM, ghN = 11)
}
\arguments{
\item{hM}{fitted Hmsc model}

\item{ghN}{order of Gauss-Hermite quadrature for approximate numerical integration}
}
\value{
the scalar WAIC
}
\description{
Computes WAIC (Widely Applicable Information Criterion) of the HMSC model
}
\details{
The result is exact for normal and probit observational models. For Poisson-type observational observational model the result is obtained through numerical integration using Gauss-Hermite quadrature.
}
\examples{
\dontrun{
WAIC = computeWAIC(hM=m)
}

}
