% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructGradient.R
\name{constructGradient}
\alias{constructGradient}
\title{hM$sc$constructGradient}
\usage{
constructGradient(hM, focalVariable, non.focalVariables = list(),
  ngrid = 20)
}
\arguments{
\item{ngrid}{}
}
\value{

}
\description{
Computes ...
}
\examples{

}
\seealso{

}
