% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToCodaObject.r
\name{convertToCodaObject}
\alias{convertToCodaObject}
\title{Hmsc$convertToCodaObject}
\usage{
convertToCodaObject(start = 1, spNamesNumbers = c(TRUE, TRUE),
  covNamesNumbers = c(TRUE, TRUE), trNamesNumbers = c(TRUE, TRUE),
  Beta = TRUE, Gamma = TRUE, V = TRUE, Sigma = TRUE, Rho = TRUE,
  Eta = TRUE, Lambda = TRUE, Alpha = TRUE, Omega = TRUE, Psi = TRUE,
  Delta = TRUE)
}
\arguments{
\item{start}{}
}
\description{
Converts Hmsc posterior structure into a named list of mcmc.list objects
}
\examples{

}
