% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralPredict.R
\name{coralPredict}
\alias{coralPredict}
\title{coralPredict}
\usage{
coralPredict(XData, XFormula, mu, V, prob = TRUE)
}
\arguments{
\item{XData}{dataframe of covariates for the sampling units to be predicted}

\item{XFormula}{a \code{\link{formula}}-class object for fixed effects}

\item{mu}{matrix of CORAL prior or posterior means}

\item{V}{matrix of CORAL prior or posterior flattened variances}

\item{prob}{whether to return predictions as probabilities (TRUE) or latent linear predictor (FALSE)}
}
\value{
Matrix of CORAL predictions with predicted sampling units in the rows and species in the columns
}
\description{
Makes predictions from CORAL models
}
