% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralSplitToBatches.R
\name{coralSplitToBatches}
\alias{coralSplitToBatches}
\title{coralSplitToBatches}
\usage{
coralSplitToBatches(Y, TrData, C.common.rare, batchN = 1)
}
\arguments{
\item{Y}{community matrix of rare species}

\item{TrData}{trait matrix (dataframe) for rare species}

\item{C.common.rare}{phylogeny similarity matrix between common and rare species}

\item{batchN}{target number of batches}
}
\value{
A named list of three elements \code{Y}, \code{TrData} and \code{C.common.rare},
each of which is a list of length \code{batchN} with splits of the corresponding input argument
}
\description{
Splits CORAL data objects that contain all rare species into smaller batches for distributed computation
}
