% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPostEstimate.R
\name{getPostEstimate}
\alias{getPostEstimate}
\title{Hmsc$getPostEstimate}
\usage{
getPostEstimate(hM, parName, r = 1, x = NULL, q = c(),
  chainIndex = 1:length(hM$postList), start = 1, thin = 1)
}
\arguments{
\item{parName}{The name of the parameter to be summarized. Can take value of
model's baseline parameters, "Omega" or "OmegaCor".}

\item{r}{Which random level calculate the parameter for. Has effect only for Eta, Lambda, Omega and OmegaCor.}

\item{q}{Which quantiles to calculate.}

\item{chainIndex}{Which posterir chains to use for summarization (defaults to all)}
}
\value{
A named list of posterior quantities. Element mean
}
\description{
Calculates mean, support and other posterior quantities for model parameter
}
\examples{
# Get posterior mean and support for species' responses to environmental covariates
postBeta = getPostEstimate(TD$m, parName='Beta')

# Get posterior mean and support for species' responses to latent factors for the first random level
postLambda = getPostEstimate(TD$m, parName='Lambda',r=1)

}
