% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMSC-package.R
\docType{package}
\name{HMSC-package}
\alias{HMSC-package}
\title{Hierarchical Modelling of Species Community}
\description{
Hierarchical Modelling of Species Community
}
\details{
\tabular{ll}{
Package: \tab HMSC\cr
Type: \tab Package\cr
Version: \tab 0.22\cr
Date: \tab 2016-10-27\cr
License: \tab GPL (>= 2)\cr
}

Implement a general framework to analyse species community data
using Bayesian hierarchical models. All estimations are carried out using MCMC.
}
\author{
F. Guillaume Blanchet \email{guillaume.blanchet@usherbrooke.ca}
}
\keyword{package}
