% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hmsc-package.R
\docType{package}
\name{Hmsc-package}
\alias{Hmsc}
\alias{Hmsc-package}
\title{Hmsc: Hierarchical Model of Species Communities}
\description{
Hierarchical Modelling of Species Communities (HMSC) is 
   a model-based approach for analyzing community ecological data. 
   This package implements it in the Bayesian framework with Gibbs MCMC sampling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.helsinki.fi/en/researchgroups/metapopulation-research-centre/hmsc}
}

}
\keyword{internal}
