% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hmsc.R
\name{Hmsc}
\alias{Hmsc}
\title{Hmsc}
\usage{
Hmsc(Y, XFormula = ~., XData = NULL, X = NULL, XScale = TRUE,
  studyDesign = NULL, ranLevels = NULL,
  ranLevelsUsed = names(ranLevels), Xs = NULL, Xv = NULL,
  TrFormula = NULL, TrData = NULL, Tr = NULL, TrScale = TRUE,
  phyloTree = NULL, C = NULL, distr = "normal", priors = NULL)
}
\arguments{
\item{Y}{matrix of species occurences or abundances}

\item{XFormula}{formula for linear regression component of HMSC}

\item{XData}{dataframe of measured covariates for formula-based specification}

\item{X}{matrix of measured covariates for direct specification}

\item{XScale}{(boolean; default is TRUE) whether to scale values in X matrix}

\item{studyDesign}{data frame of correspondence between sampling units and units on
different levels of latent factors}

\item{ranLevels}{list of HmscRandomLevel objects, specifying the structure and data for random levels}

\item{ranLevelsUsed}{vector with names of levels of latent factors that are used in the analysis, subset of colnames(dfPi)}

\item{TrFormula}{formula for linear regression dependence of Beta coefficeints on provided traits}

\item{TrData}{dataframe of measured traits for formula-based specification}

\item{Tr}{matrix of measured traits for direct specification}

\item{TrScale}{(boolean; default is TRUE) whether to scale values in the Tr matrix}

\item{distr}{string shortcut or $n_s \times 4$ matrix specifying the observation models}
}
\value{
Object of Hmsc class
}
\description{
Creates a Hmsc object
}
\examples{
Y = matrix(rnorm(100*10),100,10,dimnames=list(NULL,letters[1:10]))
X = matrix(rnorm(100*3),100,3,dimnames=list(NULL,sprintf("cov\%d",1:3)))
m = Hmsc(Y=Y, XData=as.data.frame(X), XFormula=~cov1+cov3)

}
\seealso{

}
