% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmscRandomLevel.R
\name{HmscRandomLevel}
\alias{HmscRandomLevel}
\title{Create single level of random factors in HMSC}
\usage{
HmscRandomLevel(sData = NULL, sMethod = "Full", distMat = NULL,
  xData = NULL, units = NULL, N = NULL, nNeighbours = NULL,
  sKnot = NULL, knotDist = NULL, minKnotDist = NULL)
}
\arguments{
\item{sData}{a dataframe containing spatial or temporal coordinates of units of the random level}

\item{sMethod}{a string specifying which spatial method to be used. Possible values are \code{Full}, \code{GPP} and \code{NNGP}}

\item{distMat}{a distance matrix containing the distances between units of the random level}

\item{xData}{a dataframe containing the covariates measured at the units of the random level for covariate-dependent
associations}

\item{units}{a vector, specifying the names of the units of a non-structured level}

\item{N}{number of unique units on this level}

\item{nNeighbours}{a scalar specifying the number of neighbours to be used in case the spatial method is set to \code{NNGP}. Only positive values smaller than the total number of plots are allowed.}

\item{sKnot}{a dataframe containing the knot locations to be used for the gaussian predictive process if sMethod is set to \code{GPP}}

\item{knotDist}{distance between knots used in gaussian predictive proces if sMethod is set to \code{GPP}.}

\item{minKnotDist}{minimum distance of knot to neareast point in sData}
}
\value{
a \code{HmscRandomLevel}-class object that can be used for \code{Hmsc}-class object construction
}
\description{
Specifies the structure for a single level of random factors, including whether the level is assumed to
  be spatial or not, the spatial coordinates and the potential structure of covariate-dependent nonstationarity.
}
\details{
Only one of \code{sData}, \code{distMat}, \code{xData}, \code{units} and \code{N} arguments shall be
  provided (implmentation for \code{sData} and \code{xData} is coming later).

  As a good practice we recommend to specify all available units for a random level, even if some of those are not
  used for training the model.
}
\examples{
rL = HmscRandomLevel(sData=data.frame(s1=c(1:10),s2=c(10:1)))
rL = HmscRandomLevel(units=as.factor(1:10))

}
\seealso{
[setPriors.Hmsc()]
}
