% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPosteriorFromHPC.R
\name{importPosteriorFromHPC}
\alias{importPosteriorFromHPC}
\title{importPosteriorFromHPC}
\usage{
importPosteriorFromHPC(
  m,
  postList,
  nSamples,
  thin,
  transient,
  adaptNf = rep(transient, m$nr),
  verbose = 0,
  alignPost = TRUE
)
}
\arguments{
\item{m}{a \code{Hmsc} model object}

\item{postList, nSamples, thin, transient, adaptNf, verbose, alignPost}{arguments for imported elements}
}
\description{
Computes initial parameter values before the sampling starts
}
