% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPosteriorFromHPC.R
\name{importPosteriorFromHPC}
\alias{importPosteriorFromHPC}
\title{importPosteriorFromHPC}
\usage{
importPosteriorFromHPC(
  m,
  postList,
  nSamples,
  thin,
  transient,
  alignPost = TRUE
)
}
\arguments{
\item{m}{a \code{Hmsc} model object}

\item{initPar}{a list of initial parameter values}
}
\description{
Computes initial parameter values before the sampling starts
}
