% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGamma.R
\name{plotGamma}
\alias{plotGamma}
\title{plotGamma}
\usage{
plotGamma(hM, post, param = "Gamma", trOrder = "Original",
  trVector = NULL, covOrder = "Original", covVector = NULL,
  trNamesNumbers = c(T, T), covNamesNumbers = c(T, T),
  supportLevel = 0.9, cex = c(0.8, 0.8, 0.8),
  colors = colorRampPalette(c("blue", "white", "red")),
  colorLevels = 200, mar = c(6, 9, 2, 0), smallplot = NULL,
  bigplot = NULL)
}
\arguments{
\item{post}{Posterior summary of Gamma parameters obtained from \code{\link{getPostEstimate}}}

\item{param}{Controls which parameter is plotted, current options include "Mean" for posterior mean estimate, "Support" for the level of statistical support measured by posterior probability for a positive or negative response, and "Sign" to indicate whether the response is positive, negative, or neither of these given the chosen \code{supportLevel}}

\item{trOrder}{Controls the ordering of traits, current options are "Original", and "Vector". If trOrder = "Vector", an ordering vector must be provided (see trVector)}

\item{trVector}{Controls the ordering of traits if trOrder = "Vector". If a subset of traits are listed, only those will be plotted}

\item{covOrder}{Controls the ordering of covariates, current options are "Original" and "Vector". If covOrder = "Vector", an ordering vector must be provided (see covVector)}

\item{covVector}{Controls the ordering of covariates if covOrder = "Vector". If a subset of covariates are listed, only those will be plotted}

\item{trNamesNumbers}{Logical of length 2, where first entry controls whether trait names are added to axes, and second entry controls whether traits numbers are added}

\item{covNamesNumbers}{Logical of length 2, where first entry controls whether covariate names are added to axes, and second entry controls whether covariate numbers are added}

\item{supportLevel}{Controls threshold posterior support for plotting}

\item{cex}{Controls character expansion (font size). Three values, controlling covariate names, trait names, and color legend axis labels}

\item{colors}{Controls the colors of the heatmap, default value \code{colorRampPalette(c("blue","white","red"))}}
}
\description{
Plots heatmaps of parameter estimates or posterior support values of trait effects on species' environmental responses, i.e. how environmental responses in \code{Beta} responds to covariates in \code{X}
}
