% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariancePartitioning.R
\name{plotVariancePartitioning}
\alias{plotVariancePartitioning}
\title{Hmsc$plotVariancePartitioning}
\usage{
plotVariancePartitioning(VP, ...)
}
\arguments{
\item{VP}{}
}
\description{
Plots ...
}
\examples{

}
