% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariancePartitioning.R
\name{plotVariancePartitioning}
\alias{plotVariancePartitioning}
\title{plotVariancePartitioning}
\usage{
plotVariancePartitioning(hM, VP, ...)
}
\arguments{
\item{...}{other parameters passed to the barplot function}
}
\value{

}
\description{
Plots ...
}
\examples{

}
\seealso{

}
