% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.hmsc.R
\name{predict.hmsc}
\alias{predict.hmsc}
\title{Model prediction for HMSC models}
\usage{
\method{predict}{hmsc}(object, newdata, conditional = NULL, nsample)
}
\arguments{
\item{object}{An object of the class \code{hmsc}}

\item{newdata}{An optional of class \code{HMSCdata} in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{conditional}{A character vector defining the names of the species used for the conditional prediction. If omitted, unconditional predictions are carried out. Default is NULL.}

\item{nsample}{A numerical value defining the number of samples to carry out when calculating conditional probability. If the \code{conditional} is NULL, this argument will not be considered.}

\item{\dots}{Additional arguments affecting the predictions produced.}
}
\value{
A matrix with the predicted values for each species at the sites in \code{newdata} or a matrix of the same dimension as the fitted values.
}
\description{
Obtains predictions for a model constructed using \code{hmsc}
}
\examples{

#================
### Generate data
#================
desc <- cbind(scale(1:50), scale(1:50)^2)
nspecies <- 20
commDesc <- communitySimul(X = desc, nsp = nspecies)

#=============
### Formatting
#=============
### Format data
formdata <- as.HMSCdata(Y = commDesc$data$Y, X = desc, interceptX = FALSE,
					     interceptTr = FALSE)

#==============
### Build model
#==============
modelDesc <- hmsc(formdata, niter = 2000, nburn = 1000, thin = 1, verbose = 100)

#=======================
### Calculate prediction
#=======================
predModel <- predict(modelDesc)

}
\author{
F. Guillaume Blanchet
}
\keyword{multivariate,}
\keyword{regression}
\keyword{univar,}
