% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Hmsc}
\alias{predict.Hmsc}
\title{predict.Hmsc}
\usage{
\method{predict}{Hmsc}(hM, post = poolMcmcChains(hM$postList),
  XData = NULL, X = NULL, dfPiNew = hM$dfPi, rL = hM$rL,
  Yc = NULL, mcmcStep = 1, expected = FALSE,
  predictEtaMean = FALSE)
}
\arguments{
\item{expected}{(boolean; default is FALSE)}

\item{predictEtaMean}{(boolean; default is FALSE)}
}
\value{

}
\description{
Calculates predicted values for a given hMSC model
}
\examples{

}
\seealso{

}
