% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Hmsc}
\alias{predict.Hmsc}
\title{predict.Hmsc}
\usage{
\method{predict}{Hmsc}(post, X = self$X, dfPiNew = self$dfPi,
  rL = self$rL, expected = FALSE)
}
\arguments{
\item{initPar}{initial parameters value}
}
\description{
Computes the initial values before the sampling starts
}
