% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Hmsc}
\alias{predict.Hmsc}
\title{predict.Hmsc}
\usage{
\method{predict}{Hmsc}(post = poolMcmcChains(self$postList), XData = NULL,
  X = NULL, dfPiNew = self$dfPi, rL = self$rL, expected = FALSE,
  predictEtaMean = FALSE)
}
\arguments{
\item{expected}{(boolean; default is FALSE)}

\item{predictEtaMean}{(boolean; default is FALSE)}
}
\value{

}
\description{
Calculates predicted values for a given HMSC model
}
\examples{

}
\seealso{

}
