% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictLatentFactor.R
\name{predictLatentFactor}
\alias{predictLatentFactor}
\title{predictLatentFactor}
\usage{
predictLatentFactor(unitsPred, units, postEta, postAlpha, rL,
  predictMean = FALSE)
}
\arguments{
\item{initPar}{initial parameters value}
}
\description{
Computes the initial values before the sampling starts
}
\examples{

}
