% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMcmc.R
\name{sampleMcmc}
\alias{sampleMcmc}
\title{sampleMCMC}
\usage{
sampleMcmc(samples, transient = 0, thin = 1, initPar = NULL, repN = 1,
  saveToDisk = FALSE, verbose = samples * thin/100, adaptNf = NULL,
  nChains = 1, dataParList = NULL)
}
\arguments{
\item{samples}{number of MCMC steps to be recorded}

\item{thin}{thinning between recorded MCMC samples}

\item{initPar}{initial parameters value}

\item{saveToDisk}{whether to save replicates to the disk once they are ready and discard them from RAM memory}

\item{replicates}{number of replicates the samples should be splitted}

\item{discardDataPar}{whether to discard the precomputed data-based quantities once the sampling is finished}
}
\description{
Samples the posterior with block-conditional Gibbs MCMC sampler
}
\examples{

}
