% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMcmc.R
\name{sampleMcmc}
\alias{sampleMcmc}
\title{sampleMCMC}
\usage{
sampleMcmc(hM, samples, transient = 0, thin = 1, initPar = NULL,
  verbose = samples * thin/100, adaptNf = rep(transient, hM$nr),
  nChains = 1, nCores = 1, dataParList = NULL, updater = list())
}
\arguments{
\item{samples}{number of MCMC steps to be recorded}

\item{thin}{thinning between recorded MCMC samples}

\item{initPar}{initial parameters value}

\item{verbose}{at which steps should model progress be displayed (default = samples*thin / 100)}

\item{nChains}{number of MCMC chains to run}

\item{updater}{}

\item{saveToDisk}{whether to save replicates to the disk once they are ready and discard them from RAM memory}
}
\value{

}
\description{
Samples the posterior with block-conditional Gibbs MCMC sampler
}
\examples{

}
\seealso{

}
