% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\name{setData}
\alias{setData}
\title{setData}
\usage{
setData(Y = NULL, XFormula = ~., XData = NULL, X = NULL,
  XScale = TRUE, dfPi = NULL, rL = NULL, Xs = NULL, Xv = NULL,
  TrFormula = NULL, TrData = NULL, Tr = NULL, TrScale = TRUE,
  phyloTree = NULL, C = NULL, distr = "normal", spNames = NULL,
  trNames = NULL, covNames = NULL, levelNames = NULL)
}
\arguments{
\item{Y}{matrix of species occurences or abundances}

\item{XFormula}{}

\item{X}{matrix of measured covariates}

\item{XScale}{(boolean; default is TRUE) scale values in X matrix?}

\item{dfPi}{matrix of correspondence between sampling units and units on 
different levels of latent factors}

\item{rL}{list of HmscRandomLevel objects, specifying the structure and data for random levels}

\item{TrScale}{(boolean; default is TRUE) scale values in the trait (TR) matrix?}

\item{levelNames}{}
}
\description{
Sets or resets various data elements to an Hmsc object
}
\examples{

}
