% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPriors.HmscRandomLevel.R
\name{setPriors.HmscRandomLevel}
\alias{setPriors.HmscRandomLevel}
\title{setPriors.HmscRandomLevel}
\usage{
\method{setPriors}{HmscRandomLevel}(rL, nu = NULL, a1 = NULL,
  a2 = NULL, b1 = NULL, b2 = NULL, alphapw = NULL, nfMax = NULL,
  nfMin = NULL, setDefault = FALSE, ...)
}
\arguments{
\item{rL}{a fitted \code{HmscRandomLevel} model object}

\item{nu, a1, b1, a2, b2}{parameters of the multiplicative gamma process shrinking prior}

\item{alphapw}{discrete grid prior for spatial scale parameter}

\item{nfMax}{maximum number of latent factors to be sampled}

\item{nfMin}{minimum number of latent factors to be sampled}

\item{setDefault}{logical indicating whether default priors should be used}

\item{...}{other arguments (ignored)}
}
\value{
Modified HmscRandomLevel object
}
\description{
Sets or resets priors to the Hmsc object
}
