% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPriors.R
\name{setPriors}
\alias{setPriors}
\title{setPriors}
\usage{
setPriors(...)
}
\arguments{
\item{\dots}{Hmsc or HmscRandolLevel object and other arguments.}
}
\value{
Object of same type as first input
}
\description{
Sets or resets priors to objects
}
\seealso{
setPriors.Hmsc, setPriors.HmscRandomLevel
}
