% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPriors.R
\name{setPriors}
\alias{setPriors}
\title{setPriors}
\usage{
setPriors(V0 = NULL, f0 = NULL, mGamma = NULL, UGamma = NULL,
  aSigma = NULL, bSigma = NULL, nu = NULL, a1 = NULL, b1 = NULL,
  a2 = NULL, b2 = NULL, rhopw = NULL, setDefault = FALSE)
}
\arguments{
\item{V0}{scale matrix in the Wishart prior distribution for the Beta parameters}

\item{f0}{number of degreees of freedom in the Wishart prior distribution for the Beta parameters}

\item{mGamma}{mean for the prior multivariate Gaussian distribution for Gamma parameters}

\item{UGamma}{covariance matrix for the prior multivariate Gaussian distribution for Gamma parameters}
}
\description{
Sets or resets priors to the Hmsc object
}
\examples{

}
