% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPriors.R
\name{setPriors}
\alias{setPriors}
\title{setPriors}
\usage{
setPriors(...)
}
\arguments{
\item{\dots}{Hmsc or HmscRandolLevel object and other arguments.}
}
\value{
Object of same type as first input
}
\description{
Sets or resets priors to objects
}
\examples{
# Set priors for random level so that there is minimum of 2 latent factors and maximum of 3
rL1 = HmscRandomLevel(units=TD$studyDesign$plot)
rL1 = setPriors(rL1, nfMax=3, nfMin=2)

# Set shrinkage parameters for priors of random level
rL1 = HmscRandomLevel(units=TD$studyDesign$plot)
rL1 = setPriors(rL1, a1=10, a2=10, b1=1, b2=1)

}
\seealso{
setPriors.Hmsc, setPriors.HmscRandomLevel
}
