% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateBetaLambda.R
\name{updateBetaLambda}
\alias{updateBetaLambda}
\title{updateBetaLambda}
\usage{
updateBetaLambda(Y, Z, Gamma, iV, iSigma, Eta, Psi, Delta, rho, iQg, X, Tr, Pi,
  C)
}
\description{
updates beta lambda
}
