% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variPart.R
\name{variPart}
\alias{variPart}
\title{Variation partitioning}
\usage{
variPart(hmsc, groupX)
}
\arguments{
\item{hmsc}{An object of the class \code{hmsc}.}

\item{groupX}{A vector defining how the covariates (\code{X}) are grouped. This argument is ignored if the models does not have any covariates.}
}
\value{
A list presenting how the explained variation in the data is partitioned between the different groups of the covariates (first part of thet list) and the random effects (second part of the list). Note that both autocorrelated (\code{Auto}) and non-autocorrelated (\code{Random}) are considered here. In addition, this function also calculate the amount of variation explained by traits.
}
\description{
Partition the variation of models
}
\examples{

#================
### Generate data
#================
desc <- cbind(scale(1:50), scale(1:50)^2)
nspecies <- 20
commDesc <- communitySimul(X = desc, nsp = nspecies)

#=============
### Formatting
#=============
### Format data
formdata <- as.HMSCdata(Y = commDesc$data$Y, X = desc, interceptX = FALSE, 
						   interceptTr = FALSE)

#==============
### Build model
#==============
modelDesc <- hmsc(formdata, niter = 2000, nburn = 1000, thin = 1, verbose = 100)

#==================================
### Calculate variance partitioning
#==================================
vp <- variPart(modelDesc, groupX = c("A", "B"))
vp

#===================
### Plot the results
#===================
barplot(t(vp), las = 2)

}
\author{
Gleb Tikhonov and modified by Guillaume Blanchet
}
\keyword{multivariate,}
\keyword{regression}
\keyword{univar,}
